Learning about using protocols.
Aside from learning how to work with protocols. Making DevUtil.m reminded me of making a simple PHP file to be included and have a specific utility task. This is something we don't do in Java or Kotlin. As we are required to use classes, objects in Kotlin, but in the same way we can do this also in Javascript.